#78: Refactoring.
